requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
requests==2.26.0
discord.py==2.0.1
eventlet==0.33.0